// 07 - Crypto Dashboard Projekt/src/components/types.ts
export type Coin = {
    id: string;
    symbol: string;
    name: string;
    image: string;
    current_price: number;
    price_change_24h: number;
    price_change_percentage_24h: number;
    market_cap: number;
    market_cap_rank: number;
    market_cap_change_24h: number;
    market_cap_change_percentage_24h: number;
    total_volume: number;
    high_24h: number;
    low_24h: number;
    ath: number;
    ath_change_percentage: number;
    ath_date: string;
    atl: number;
    atl_change_percentage: number;
    atl_date: string;
    fully_diluted_valuation: number | null;
    total_supply: number | null;
    max_supply: number | null;
    circulating_supply: number;
    last_updated: string;
    roi: any | null;
};